[Refactor] 모든 component 위치 리팩토링
- 다른 팀원들과 코드가 병합될때 이상하게 겹치거나 헷갈리는 것을 방지
