+++ sidebar/src/components/UI/molecules/DetailListItem/DetailListItem.js
@@ -3,10 +3,9 @@ import styled from 'styled-components';
 import { connect } from 'react-redux';
 import RemoveCircleIcon from '@material-ui/icons/RemoveCircle';
 import DateRangeIcon from '@material-ui/icons/DateRange';
-import Span from '../atoms/Span';
-import StyledButton from '../atoms/StyledButton';
-import DatePicker from '../atoms/DatePicker';
-import { removeDetailAction } from '../../../modules/actions/createbucket';
+import Span from '../../atoms/Span/Span';
+import StyledButton from '../../atoms/StyledButton/StyledButton';
+import { removeDetailAction } from '../../../../modules/actions/createbucket';
 
 const DetailListItemWrapper = styled.div`
   display: flex;
