--- sidebar/src/components/pages/BucketCreatePage.js
@@ -1,8 +1,0 @@
-import React, { useState } from 'react';
-import BucketCreateTemplate from '../templates/BucketCreateTemplate';
-
-const BucketCreatePage = () => {
-  return <BucketCreateTemplate />;
-};
-
-export default BucketCreatePage;
