+++ sidebar/src/components/pages/BucketCreatePage/BucketCreatePage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import BucketCreateTemplate from '../../templates/BucketCreateTemplate/BucketCreateTemplate';
+
+const BucketCreatePage = () => {
+  return <BucketCreateTemplate />;
+};
+
+export default BucketCreatePage;
