+++ sidebar/src/components/templates/BucketCreateTemplate/BucketCreateTemplate.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
-import BucketSearchBar from '../UI/molecules/BucketSearchBar';
-import BucketContents from '../UI/organisms/BucketContents';
-import DetailList from '../UI/organisms/DetailList';
-import Title from '../UI/molecules/Title';
-import Buttons from '../UI/organisms/Buttons';
+import BucketSearchBar from '../../UI/molecules/BucketSearchBar/BucketSearchBar';
+import BucketContents from '../../UI/organisms/BucketContents/BucketContents';
+import DetailList from '../../UI/organisms/DetailList/DetailList';
+import Title from '../../UI/molecules/Title/Title';
+import Buttons from '../../UI/organisms/Buttons/Buttons';
 
 const BucketCreateTemplateWrapper = styled.div`
   display: flex;
